+++ bank-feeds/Codat/BankFeeds/Utils/Constants.cs
@@ -22,12 +22,12 @@ namespace Codat.BankFeeds.Utils
         /// <summary>
         /// The version of the SDK.
         /// </summary>
-        public const string SdkVersion = "8.0.0";
+        public const string SdkVersion = "8.0.1";
 
         /// <summary>
         /// The version of the SDK generator used to create this SDK.
         /// </summary>
-        public const string SdkGenVersion = "2.799.0";
+        public const string SdkGenVersion = "2.879.6";
 
         /// <summary>
         /// The version of the OpenAPI document used to generate this SDK.